refactor(OutOfStockError): entity layer로 분리
