+++ src/refactoring/entities/product/model/errors.ts
@@ -1,5 +1,6 @@
 import {
   CreateProductInput,
+  Product,
   REQUIRED_CREATE_PRODUCT_PROPERTIES,
 } from "./product";
 
@@ -37,5 +38,13 @@ export class DuplicatedProductNameError extends Error {
     const message = `The product name "${productName}" is already in use.`;
     super(message);
     this.name = DuplicatedProductNameError.NAME;
+  }
+}
+export class OutOfStockError extends Error {
+  static readonly name = "OutOfStockError";
+
+  constructor(product: Product) {
+    super(`${product.name} is out of stock`);
+    this.name = OutOfStockError.name;
   }
 }
